Legal move checker for pawns
We have added a legal move checker using the rules for pawns
(currently applied to all pieces). The function `getLegalMoves`
returns a vector of pairs expressing the coordinates of where a piece
may move to given its current location and the game state. Besides
verifying legal moves, this will also be used to verify check and
checkmate as well as to implement an AI.