Legal move checking for all pieces
We have three different cases: (1) pawns, (2) kings and knights and (3) queens, rooks and bishops.

Pawns have a lot of special rules, so we handle them separately.

For kings and knights, there's a predefined set of spaces that are an absolute distance from the current
location. We loop over that list.

For queens, rooks and bishops, there's a defined set of directions the pieces may move in, and they may continue
until the square before occupied by a piece of the same color, until we've hit a square occupied by an enemy,
or we've hit the boundary of the board.